separate shield configurations
