+++ ios/shieldConfigurator/ShieldConfigurationExtension.swift
@@ -14,27 +14,6 @@ import OSLog
 import DeviceActivity
 import FamilyControls
 
-extension UIColor {
-    convenience init(hex: String, alpha: CGFloat = 1.0) {
-        var hexFormatted: String = hex.trimmingCharacters(in: .whitespacesAndNewlines).uppercased()
-        
-        if hexFormatted.hasPrefix("#") {
-            hexFormatted.remove(at: hexFormatted.startIndex)
-        }
-        
-        assert(hexFormatted.count == 6, "Hex color string must be 6 characters long")
-        
-        var rgbValue: UInt64 = 0
-        Scanner(string: hexFormatted).scanHexInt64(&rgbValue)
-        
-        let red = CGFloat((rgbValue & 0xFF0000) >> 16) / 255.0
-        let green = CGFloat((rgbValue & 0x00FF00) >> 8) / 255.0
-        let blue = CGFloat(rgbValue & 0x0000FF) / 255.0
-        
-        self.init(red: red, green: green, blue: blue, alpha: 1)
-    }
-}
-
 // Override the functions below to customize the shields used in various situations.
 // The system provides a default appearance for any methods that your subclass doesn't override.
 // Make sure that your class name matches the NSExtensionPrincipalClass in your Info.plist.
@@ -43,56 +22,6 @@ class ShieldConfigurationExtension: ShieldConfigurationDataSource {
     
   private var eventModel: AppEvent?
   private var logger = Logger()
-  
-  func limitShield(
-    applicationName: String = "",
-    eventName: String = "",
-    impulseTime: Int = 0,
-    openLimite: String = "",
-    opens: Int = 0,
-    shieldButtonEnable: Bool = true
-  ) -> ShieldConfiguration{
-    
-    let secondaryButtonText = "Continuar en \(impulseTime) seg"
-    
-    let disableButtonColorBack = UIColor(hex: "#e6e6e6")
-    let disableButtonColotText = UIColor.black
-    
-    let enableButtonColorBack = UIColor.black
-    let enableButtonColorText = UIColor(hex: "#FDE047")
-    
-    let buttonBackground = shieldButtonEnable ? enableButtonColorBack : disableButtonColorBack
-    let buttonTextColor = shieldButtonEnable ? enableButtonColorText : disableButtonColotText
-    
-    let numberOfOpenLimite = Int(openLimite)
-  
-    let openLimiteText = numberOfOpenLimite ?? 0 > 0 ? "\(opens)/\(numberOfOpenLimite ?? 0)" : "\(opens)"
-    let subtitle = "Tienes configurado bloquear \(applicationName) durante \(eventName) \n\n\n Intentos de apertura: \(openLimiteText)"
-    
-    return ShieldConfiguration(
-      backgroundBlurStyle: UIBlurEffect.Style.light,
-      backgroundColor: UIColor(hex: "#FDE047"),
-      icon: UIImage(named: "impulse-icon"),
-      title: ShieldConfiguration.Label(text: "\n\n¿Quieres\ncontinuar?", color: UIColor.black),
-      subtitle: ShieldConfiguration.Label(text: subtitle, color: UIColor.black),
-      primaryButtonLabel: ShieldConfiguration.Label(text: secondaryButtonText, color: enableButtonColorText),
-      primaryButtonBackgroundColor: enableButtonColorBack,
-      secondaryButtonLabel: ShieldConfiguration.Label(text: "Cerrar App", color: UIColor.black)
-    )
-    
-  }
-  
-  func blockShield(applicationName: String = "", eventName: String = "") -> ShieldConfiguration{
-    return ShieldConfiguration(
-      backgroundBlurStyle: UIBlurEffect.Style.light,
-      backgroundColor: UIColor(hex: "#FDE047"),
-      icon: UIImage(named: "lock-shield-icon"),
-      title: ShieldConfiguration.Label(text: "\(applicationName) esta\nbloqueada", color: UIColor.black),
-      subtitle: ShieldConfiguration.Label(text: "Tienes configurado bloquear \(applicationName) durante \(eventName)", color: UIColor.black),
-      primaryButtonLabel: ShieldConfiguration.Label(text: "Cerrar", color: UIColor.black),
-      primaryButtonBackgroundColor: UIColor.white
-    )
-  }
   
   override func configuration(shielding application: Application) -> ShieldConfiguration {
     
@@ -114,7 +43,7 @@ class ShieldConfigurationExtension: ShieldConfigurationDataSource {
           let blockName = shieldConfigurationData["blockName"] as? String ?? ""
           
           logger.info("Impulse: shield for block  \(blockName, privacy: .public)")
-          return blockShield(applicationName: application.localizedDisplayName ?? "", eventName: blockName)
+          return ShieldConfigurationBlock.blockShield(applicationName: application.localizedDisplayName ?? "", eventName: blockName)
         }
       }
             
@@ -133,7 +62,7 @@ class ShieldConfigurationExtension: ShieldConfigurationDataSource {
           logger.info("Impulse: configure shield with primary button \(shieldButtonEnable)")
           
           // Show limit shield
-          return limitShield(
+          return ShieldConfigurationLimit.limitShield(
             applicationName: application.localizedDisplayName ?? "",
             eventName: limitName,
             impulseTime: impulseTime,
@@ -179,7 +108,7 @@ class ShieldConfigurationExtension: ShieldConfigurationDataSource {
             
             if type == "block" {
               logger.info("Shield for block  \(type, privacy: .public)")
-              return blockShield(applicationName: webDomain.domain ?? "", eventName: blockName)
+              return ShieldConfigurationBlock.blockShield(applicationName: webDomain.domain ?? "", eventName: blockName)
             }
           }
         }
@@ -201,7 +130,7 @@ class ShieldConfigurationExtension: ShieldConfigurationDataSource {
             
             if type == "limit" && isUsageWarning {
               // Show limit shield
-              return limitShield(
+              return ShieldConfigurationLimit.limitShield(
                 applicationName: webDomain.domain ?? "",
                 eventName: limitName,
                 impulseTime: impulseTime,
