+++ ios/shieldConfigurator/UIColorExtension.swift
@@ -1,0 +1,24 @@
+
+import Foundation
+import UIKit
+
+extension UIColor {
+  convenience init(hex: String, alpha: CGFloat = 1.0) {
+    var hexFormatted: String = hex.trimmingCharacters(in: .whitespacesAndNewlines).uppercased()
+    
+    if hexFormatted.hasPrefix("#") {
+        hexFormatted.remove(at: hexFormatted.startIndex)
+    }
+    
+    assert(hexFormatted.count == 6, "Hex color string must be 6 characters long")
+    
+    var rgbValue: UInt64 = 0
+    Scanner(string: hexFormatted).scanHexInt64(&rgbValue)
+    
+    let red = CGFloat((rgbValue & 0xFF0000) >> 16) / 255.0
+    let green = CGFloat((rgbValue & 0x00FF00) >> 8) / 255.0
+    let blue = CGFloat(rgbValue & 0x0000FF) / 255.0
+    
+    self.init(red: red, green: green, blue: blue, alpha: 1)
+  }
+}
